backend/backend/backend/requirements-production.txt
# Production Requirements for RichesReach Backend
# Core Django
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.2
# GraphQL
graphene-django==3.2.3
django-graphql-jwt==0.4.0
# WebSocket and Channels
channels==4.0.0
channels-redis==4.1.0
# Celery
celery==5.3.0
django-celery-results==2.5.0
redis==4.6.0
# Database
psycopg2-binary==2.9.9
# Production Server
gunicorn==22.0.0
whitenoise==6.6.0
# AI/ML
openai==1.106.0
pandas==2.3.2
scikit-learn==1.7.1
joblib==1.5.2
numpy==2.3.2
# Stock Data
requests==2.32.3
yfinance==0.2.28
# Environment
python-decouple==3.8
dj-database-url==2.2.0
# Monitoring
sentry-sdk[django]==2.18.0
# Security
django-environ==0.11.2
# Email
django-anymail==10.2
# Caching
django-redis==5.4.0
# File Storage (if using S3)
boto3==1.35.0
django-storages==1.14.2
